Added globals, chunk rendering functional, graphical info display

+++ src/CMD_VertexTypes.h
@@ -23,7 +23,7 @@ GPI_VertexLayout CMD_GetChunkVertexLayout();
 
 CMD_ChunckVertex CMD_MapChunkVertexData(
     uint8_t posX, // pos is relative to chunckPos 
-    uint8_t posY, 
+    uint16_t posY, 
     uint8_t posZ,
     uint8_t textureKey, // 0 - 3 key (which part of quad is this)
     uint32_t textureIndex // index on 16*16 texture atlas
